# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  bn-mobile_server_internal_delivery_data_request_authReq.GoogleLoginRequest:
    properties:
      idToken:
        type: string
    required:
    - idToken
    type: object
  bn-mobile_server_internal_delivery_data_response.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: API backend Belajar Ngaji (Go + Gin).
  title: BN Mobile API
  version: "1.0"
paths:
...
  /auth/google:
    post:
      consumes:
      - application/json
      description: Validate Google ID token and return access token with refresh-token
        cookie.
      parameters:
      - description: Google login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/bn-mobile_server_internal_delivery_data_request_authReq.GoogleLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
      summary: Login with Google ID token
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke refresh token and clear refresh-token cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
      summary: Logout current session
      tags:
      - Auth
  /auth/me:
    get:
      description: Return current user profile from Bearer token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get current authenticated user
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Issue a new access token and rotate refresh-token cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bn-mobile_server_internal_delivery_data_response.APIResponse'
      summary: Refresh access token
      tags:
      - Auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Bearer access token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"